Feature: add range-map module, refactor code by review
